requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-environ==0.11.2
django-ninja==1.1.0
dnspython==2.6.1
email_validator==2.2.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
ninja-schema==0.13.6
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
sqlparse==0.5.0
typing_extensions==4.12.2
uvicorn==0.30.1
